PSMEL-1250- gmail one click de authroization

+++ Postman/Wizard/NewWizard.php
@@ -1056,6 +1056,7 @@ class Post_SMTP_New_Wizard {
                 [
                     '_wpnonce' => $nonce,
                     'action' => 'remove_oauth_action',
+                    'id' => $id,
                 ],
                 admin_url( 'admin-post.php' )
             ) );
@@ -2438,15 +2439,38 @@ class Post_SMTP_New_Wizard {
             wp_die( esc_html__( 'Nonce verification failed. Please try again.', 'post-smtp' ) );
         }
 
-        // Remove the OAuth access token option from the WordPress database.
-        delete_option( 'postman_auth_token' );
+        $redirect_url = admin_url( "admin.php?page=postman/configuration_wizard" );
 
-        // Redirect the user back to the settings page with a success query parameter.
-        wp_redirect( admin_url( "admin.php?socket=gmail_api&step=2&page=postman/configuration_wizard" ) );
+        if ( $this->existing_db_version == POST_SMTP_DB_VERSION ) {
+            $id = isset( $_GET['id'] ) ? absint( $_GET['id'] ) : 0;
+            if ( $id ) {
+                $mail_connections = get_option( 'postman_connections', array() );
 
-        // Terminate script execution to prevent further processing after the redirect.
+                if ( isset( $mail_connections[ $id ] ) && isset( $mail_connections[ $id ]['provider'] ) && $mail_connections[ $id ]['provider'] === 'gmail_api' ) {
+                  // Reset sensitive tokens to empty values
+                    $mail_connections[ $id ]['access_token']        = '';
+                    $mail_connections[ $id ]['refresh_token']       = '';
+                    $mail_connections[ $id ]['token_expires']       = '';
+                    $mail_connections[ $id ]['auth_token_expires']  = '';
+					$mail_connections[ $id ]['sender_email']  = '';
+					$mail_connections[ $id ]['sender_name']  = '';
+                 
+                    update_option( 'postman_connections', $mail_connections );
+
+                    // Redirect back to the same Gmail mailer ID wizard
+                    $redirect_url = admin_url( "admin.php?socket=gmail_api&id={$id}&step=2&page=postman/configuration_wizard" );
+                }
+            }
+        } else {
+            delete_option( 'postman_auth_token' );
+            $redirect_url = admin_url( "admin.php?socket=gmail_api&step=2&page=postman/configuration_wizard" );
+        }
+
+        // Redirect the user back with success
+        wp_redirect( $redirect_url );
         exit;
     }
+
 
     /**
      * Handles the OAuth redirect, retrieves the token parameters from the URL,
